fix(consensus): remove env-derived min-parents rule; deterministic blue-set truncation; saturating blue-score; checked coinbase split
Audit fixes (fork/overflow prevention):
- A1 CRIT: SelfishMiningGuard min-DAG-parents was read from SHADOWDAG_MIN_DAG_PARENTS/
  NETWORK env -> nodes could disagree on validity -> chain split. Now a pure
  function of height + MIN_DAG_PARENTS constant.
- A2 HIGH: blue-set truncation iterated a HashSet (non-deterministic) -> nodes kept
  different subsets. Sort before truncate.
- A3 HIGH: blue_score/chain_height used unchecked + -> wrap on corrupt stored score.
  saturating_add.
- A4 HIGH: coinbase split used unchecked multiply. u128 intermediate.

+++ engine/consensus/validation/block_validator.rs
@@ -983,7 +983,10 @@ impl BlockValidator {
         // emission must be split correctly.
         {
             use crate::config::genesis::genesis::{DEV_REWARD_PCT, MINER_REWARD_PCT};
-            let expected_miner_base = (expected_reward * MINER_REWARD_PCT) / 100;
+            // u128 intermediate: avoid overflow on the multiply (consensus check
+            // on attacker-influenced height), matching rewards/reward.rs.
+            let expected_miner_base =
+                ((expected_reward as u128 * MINER_REWARD_PCT as u128) / 100) as u64;
             let expected_dev_base = expected_reward - expected_miner_base;
             // Dev share must be at least the expected amount
             if cb.outputs[1].amount < expected_dev_base {
